+++ frontend/src/pages/Login/ForgotPassword/index.js
@@ -1,0 +1,1 @@
+export * from './ForgotPassword'
