+++ frontend/src/pages/Login/index.js
@@ -1,6 +1,7 @@
 export * from './Login/Login'
 export * from './AuthLayout'
-export * from './ForgotPassword/ForgotPassword'
+export * from './ForgotPassword'
 export * from './Register/Register'
 export * from './Login.styles'
 export * from './components/SectionLogin'
+export * from './ResetPassword'
